+++ src/main/java/stsjorbsmod/powers/MindGlassPower.java
@@ -7,7 +7,10 @@ import com.megacrit.cardcrawl.cards.DamageInfo;
 import com.megacrit.cardcrawl.core.AbstractCreature;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.powers.AbstractPower;
+import stsjorbsmod.actions.PostAoeDamageStatsAction;
+import stsjorbsmod.actions.PreAoeDamageStatsAction;
 import stsjorbsmod.memories.OnModifyMemoriesSubscriber;
+import stsjorbsmod.relics.MindGlassRelic;
 
 public class MindGlassPower extends CustomJorbsModPower implements OnModifyMemoriesSubscriber {
     public static final StaticPowerInfo STATIC = StaticPowerInfo.Load(MindGlassPower.class);
@@ -26,14 +29,21 @@ public class MindGlassPower extends CustomJorbsModPower implements OnModifyMemor
 
     @Override
     public void onGainClarity(String id) {
-        AbstractDungeon.actionManager.addToBottom(
-                new DamageAllEnemiesAction(
-                        null,
-                        DamageInfo.createDamageMatrix(damage, true),
-                        DamageInfo.DamageType.NORMAL,
-                        // TODO: More impactful and relevant FX. See FlashAtkImgEffect.loadImage() and
-                        //  FlashAtkImgEffect.playSound() for usage of AttackEffect in base game.
-                        AbstractGameAction.AttackEffect.BLUNT_HEAVY));
+        MindGlassRelic relic = (MindGlassRelic) AbstractDungeon.player.getRelic(MindGlassRelic.ID);
+        if (relic != null) {
+            PreAoeDamageStatsAction preAction = new PreAoeDamageStatsAction();
+            PostAoeDamageStatsAction postAction = new PostAoeDamageStatsAction(relic, preAction);
+            addToBot(preAction);
+            addToBot(
+                    new DamageAllEnemiesAction(
+                            null,
+                            DamageInfo.createDamageMatrix(damage, true),
+                            DamageInfo.DamageType.NORMAL,
+                            // TODO: More impactful and relevant FX. See FlashAtkImgEffect.loadImage() and
+                            //  FlashAtkImgEffect.playSound() for usage of AttackEffect in base game.
+                            AbstractGameAction.AttackEffect.BLUNT_HEAVY));
+            addToBot(postAction);
+        }
         AbstractDungeon.actionManager.addToBottom(
                 new RemoveSpecificPowerAction(this.owner, this.owner, MindGlassPower.POWER_ID));
     }